ethereum-backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethereum-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anvil>=0.0.2",
    "ape-alchemy>=0.8.10",
    "ape-foundry>=0.8.10",
    "ape-infura>=0.8.4",
    "ape-solidity>=0.8.5",
    "bcrypt==3.2.0",
    "cryptography>=46.0.3",
    "eth-ape>=0.8.41",
    "fastapi>=0.123.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest-mock>=3.15.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "setuptools>=80.9.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
]
